// src/autogen/interfaces/anthology/2022/03/25/SchemaTaggedPayloadProtocol.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SchemaTaggedPayload {
  accessControlPolicy?: string;
  applicationVersion?: string | number;
  batch?: string;
  createdAt?: string | number;
  creator?: string;
  data: any;
  /**
   * checksum value derived from a serialized representation of the .data attribute; this is logic dependent and is for verification by the same logic
   */
  dataChecksum?: string;
  device?: string;
  fingerprint?: string;
  id?: string | number;
  isPreprocessed?: boolean;
  name?: string;
  project?: string;
  /**
   * version of the Tagged Payload protocol (this particular protocol is version 2022-03-25.1)
   */
  protocolVersion: string | number;
  schemaName: string;
  schemaUrl?: string;
  schemaVersion: string | number;
  [k: string]: unknown;
}