// src/autogen/interfaces/SchemaTaggedPayloadJsonSchema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * json schema for the schema SchemaTaggedPayload trasnformation objects. The schema for the storage of their scehams. Yes, schema of schema
 */
export interface SchemaTaggedPayloadJsonSchemaSchema {
  title: string;
  version: string;
  [k: string]: unknown;
}